copies operators.py into module
